+++ Vidly/ViewModel/CustomerFormViewModel.cs
@@ -1,8 +1,9 @@
 using System.Collections.Generic;
+using Vidly.Models;
 
-namespace Vidly.Models
+namespace Vidly.ViewModel
 {
-    public class NewCustomerViewModel
+    public class CustomerFormViewModel
     {
         public IEnumerable<MembershipType> MembershipTypes { get; set; }
         public Customer Customer { get; set; }
