+++ Vidly/Views/Customers/CustomerForm.cshtml
@@ -1,4 +1,4 @@
-@model NewCustomerViewModel
+@model Vidly.ViewModel.CustomerFormViewModel
 
 @{
     ViewBag.Title = "New";
